Add transform plotting

+++ schlam/cpp/src/tft/transformable_3d.cpp
@@ -5,7 +5,7 @@
 
 namespace tft {
 
-Transformable3D::Transformable3D(const Eigen::Vector3f &aVector,
+Transformable3D::Transformable3D(const Eigen::Vector3d &aVector,
                                  const std::string &aCoordinateFrame)
     : vector(aVector), coordinateFrame(aCoordinateFrame) {}
